Moved MetadataResultImpl to core.
