+++ gui/src/main/java/org/verapdf/gui/ValidateWorker.java
@@ -28,10 +28,10 @@ import org.verapdf.metadata.fixer.impl.pb.FixerConfigImpl;
 import org.verapdf.metadata.fixer.utils.FileGenerator;
 import org.verapdf.metadata.fixer.utils.FixerConfig;
 import org.verapdf.model.ModelParser;
-import org.verapdf.pdfa.MetadataFixerResult;
-import org.verapdf.pdfa.MetadataFixerResult.RepairStatus;
 import org.verapdf.pdfa.PDFAValidator;
+import org.verapdf.pdfa.results.MetadataFixerResult;
 import org.verapdf.pdfa.results.ValidationResult;
+import org.verapdf.pdfa.results.MetadataFixerResult.RepairStatus;
 import org.verapdf.pdfa.validation.ValidationProfile;
 import org.verapdf.pdfa.validators.Validators;
 import org.verapdf.report.HTMLReport;
