+++ metadata-fixer/src/main/java/org/verapdf/metadata/fixer/entity/Metadata.java
@@ -1,11 +1,11 @@
 package org.verapdf.metadata.fixer.entity;
 
 import org.verapdf.metadata.fixer.impl.MetadataFixerImpl;
-import org.verapdf.metadata.fixer.impl.MetadataFixerResultImpl;
 import org.verapdf.metadata.fixer.schemas.AdobePDF;
 import org.verapdf.metadata.fixer.schemas.DublinCore;
 import org.verapdf.metadata.fixer.schemas.XMPBasic;
 import org.verapdf.pdfa.flavours.PDFAFlavour;
+import org.verapdf.pdfa.results.MetadataFixerResultImpl;
 
 /**
  * Current interface provide necessary behavior of pdf metadata
