+++ metadata-fixer/src/main/java/org/verapdf/metadata/fixer/entity/PDFDocument.java
@@ -3,8 +3,8 @@ package org.verapdf.metadata.fixer.entity;
 import java.io.OutputStream;
 
 import org.verapdf.metadata.fixer.impl.MetadataFixerImpl;
-import org.verapdf.metadata.fixer.impl.MetadataFixerResultImpl;
-import org.verapdf.pdfa.MetadataFixerResult;
+import org.verapdf.pdfa.results.MetadataFixerResult;
+import org.verapdf.pdfa.results.MetadataFixerResultImpl;
 
 /**
  * Current interface provide necessary behavior of pdf document
@@ -46,15 +46,15 @@ public interface PDFDocument {
 	 * In {@link MetadataFixerResultImpl} must set 1 of 3 states:
 	 * <ul>
 	 * <li>
-	 * {@link org.verapdf.pdfa.MetadataFixerResult.RepairStatus#FIX_ERROR}
+	 * {@link org.verapdf.pdfa.results.MetadataFixerResult.RepairStatus#FIX_ERROR}
 	 * if got problems with document save
 	 * </li>
 	 * <li>
-	 * {@link org.verapdf.pdfa.MetadataFixerResult.RepairStatus#NO_ACTION}
+	 * {@link org.verapdf.pdfa.results.MetadataFixerResult.RepairStatus#NO_ACTION}
 	 * if metadata and information dictionary was not changed
 	 * </li>
 	 * <li>
-	 * {@link org.verapdf.pdfa.MetadataFixerResult.RepairStatus#SUCCESS}
+	 * {@link org.verapdf.pdfa.results.MetadataFixerResult.RepairStatus#SUCCESS}
 	 * if document save successful
 	 * </li>
 	 * </ul>
