+++ metadata-fixer/src/main/java/org/verapdf/metadata/fixer/impl/MetadataFixerImpl.java
@@ -13,8 +13,9 @@ import org.verapdf.metadata.fixer.utils.FixerConfig;
 import org.verapdf.metadata.fixer.utils.ProcessedObjectsInspector;
 import org.verapdf.metadata.fixer.utils.ValidationStatus;
 import org.verapdf.pdfa.MetadataFixer;
-import org.verapdf.pdfa.MetadataFixerResult;
 import org.verapdf.pdfa.flavours.PDFAFlavour;
+import org.verapdf.pdfa.results.MetadataFixerResult;
+import org.verapdf.pdfa.results.MetadataFixerResultImpl;
 import org.verapdf.pdfa.results.ValidationResult;
 import org.verapdf.pdfa.validation.ProfileDirectory;
 import org.verapdf.pdfa.validation.Profiles;
