+++ metadata-fixer/src/main/java/org/verapdf/metadata/fixer/impl/pb/model/MetadataImpl.java
@@ -14,7 +14,6 @@ import org.apache.xmpbox.type.BadFieldValueException;
 import org.apache.xmpbox.xml.XmpSerializer;
 import org.verapdf.metadata.fixer.entity.InfoDictionary;
 import org.verapdf.metadata.fixer.entity.Metadata;
-import org.verapdf.metadata.fixer.impl.MetadataFixerResultImpl;
 import org.verapdf.metadata.fixer.impl.pb.schemas.AdobePDFSchemaImpl;
 import org.verapdf.metadata.fixer.impl.pb.schemas.DublinCoreSchemaImpl;
 import org.verapdf.metadata.fixer.impl.pb.schemas.XMPBasicSchemaImpl;
@@ -22,6 +21,7 @@ import org.verapdf.metadata.fixer.schemas.AdobePDF;
 import org.verapdf.metadata.fixer.schemas.DublinCore;
 import org.verapdf.metadata.fixer.schemas.XMPBasic;
 import org.verapdf.pdfa.flavours.PDFAFlavour;
+import org.verapdf.pdfa.results.MetadataFixerResultImpl;
 
 import java.io.IOException;
 import java.io.OutputStream;
